Support specifying ranges of alternate values
For simplicity, we require both ends of the range to be specified an
literals.

This only works if the `complex-expressions` feature is enabled, as
`syn` can't parse ranges without its `full` feature being enabled, which
significantly increases compile times. We give a useful error if you
attempt to use ranges without this feature being enabled.