Support specifying ranges of alternate values (#107)
For simplicity, we require both ends of the range to be specified an
literals.

This only works if the `complex-expressions` feature is enabled, as
`syn` can't parse ranges without its `full` feature being enabled, which
significantly increases compile times. We give a useful error if you
attempt to use ranges without this feature being enabled.

* Bump minimum rustversion number

Pre-1.0.4 it pulled in syn with the "full" feature, which broke some of
our tests for testing when full is disabled, in minimal versions mode.