Bump trybuild to 1.0.98, and MSRV to 1.70.0
This necessitates bumping MSRV to 1.70.0, which is a significant change
from 1.56.0.

1.70.0 was more than a year ago, and seems like a reasonable bump, but
it's a bit unfortunate that we're doing this bump only for a test-only
dependency.

As we're not actually changing any production code here, things are
likely to continue to work with older rust versions, but we will not
continue testing this, and make no promises versions older than 1.70.0
will keep working.